refactor: update import and css override
